% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asp.R
\name{aspinternal}
\alias{aspinternal}
\title{Title}
\usage{
aspinternal(t0, Time, Status, Z, TRT)
}
\arguments{
\item{t0}{- pre-specified time point of analysis}

\item{Time}{- Observed imes}

\item{Status}{- Censoring status (0 = censored, 1 = observed)}

\item{Z}{- Non-treatment group covariates}

\item{TRT}{- Treatment group indicator (0 = control, 1 = treatment)}
}
\value{
\itemize{
  \item SPD - adjusted survival probability difference estimate
  \item SED - standard error estimate of adjusted survival probability difference estimate
  \item S0 - adjusted survival probability of treatment group 0
  \item S1 - adjusted survival probability of treatment group 1
  \item se10 - standard error of adjusted survival probability of treatment group 0
  \item se11 - standard error of adjusted survival probability of treatment group 1
}
}
\description{
Title
}
\keyword{internal}
