% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspbinary.R
\name{ESaspbinary}
\alias{ESaspbinary}
\title{Effect size calculation for binary covariates}
\usage{
ESaspbinary(
  alpha0,
  alpha1,
  gamma0,
  beta2,
  crate,
  t0,
  maxE,
  n,
  NN,
  alpha = 0.05,
  beta = 0.2,
  max.iter,
  p
)
}
\arguments{
\item{alpha0}{- Parameter to specify in Weibull model. See Details for more information.}

\item{alpha1}{- Parameter to specify in Weibull model. See Details for more information. \eqn{\alpha_1  = 0 } means there are proportional hazards; \eqn{\alpha_1 \neq 0 }  means the proportional hazards assumption is violated}

\item{gamma0}{- Parameter to specify in Weibull model. See Details for more information.}

\item{beta2}{- Vector of coefficients for binary covariates}

\item{crate}{- Censoring rate, assumes an exponential distribution}

\item{t0}{- Pre-specified time at which adjusted SPs for each group are calculated}

\item{maxE}{- Maximum enrollment time. Assumes uniform enrollment between [0, E]}

\item{n}{- Sample size per group}

\item{NN}{- Number of iterations used to calculate the maximum information}

\item{alpha}{- Targeted type I error rate}

\item{beta}{- Targeted type II error rate}

\item{max.iter}{- Maximum number of iterations to calculate the effect size}

\item{p}{- Vector of probabilities for binary covariates}
}
\description{
Effect size calculation for binary covariates
}
\details{
See Details section in \code{\link{Imaxasp}} on how trial data are simulated.
}
\examples{
\dontrun{
ESaspbinary(alpha0 = 1.5, alpha1 = -1, gamma0 = -log(0.4), beta2=0, crate=0, t0=1,
 maxE=2, n=212, NN = 10000, alpha=0.05, beta = 0.2, max.iter=10)
}
}
\references{
Zhang, P.K., Logan, B.L., and Martens, M.J. (2024). Covariate-adjusted Group Sequential Comparisons of Survival Probabilities. \emph{arXiv}

Zhang, X., Loberiza, F. R., Klein, J. P., and Zhang, M.-J. (2007). A SAS Macro for
Estimation of Direct Adjusted Survival Curves Based on a Stratified Cox Regression
Model. \emph{Comput Methods Programs Biomed} \strong{88(2)}, 95–101.

Zucker, D.M. (1998) Restricted Mean Life with Covariates: Modification and Extension
of a Useful Survival Analysis Method. \emph{J Am Stat Assoc} \strong{93(442)}, 702-709
}
