% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsRMST.R
\name{gsRMST}
\alias{gsRMST}
\title{Group sequential test of adjusted restricted mean survival times}
\usage{
gsRMST(tau, Time, Status, Z, TRT, E, Imax, alpha = 0.05, u)
}
\arguments{
\item{tau}{Pre-specified survival horizon time for adjusted restricted mean survival times}

\item{Time}{Observed times}

\item{Status}{Event indicator (0 = censored, 1 = observed)}

\item{Z}{Non-treatment group covariates}

\item{TRT}{Treatment group indicator (0 = Control, 1 = Treatment)}

\item{E}{Enrollment times}

\item{Imax}{Maximum information}

\item{alpha}{Targeted type I error rate}

\item{u}{Calendar times of analysis}
}
\value{
\itemize{
 \item  muD - Vector of estimated restricted mean survival time differences
  \item SED - Vector of estimated standard errors of estimated restricted mean survival time differences
  \item IF - Vector of information fractions
  \item TStat Vector of test statistics
  \item Crit Vector of critical values
}
}
\description{
Performs a group sequential test of adjusted restricted mean survival times (RMST)s between two treatment groups
when provided the time point of analysis, the observed times, event indicator,
covariates, treatment group indicator, enrollment times, and calendar times of analysis.
}
\examples{
\dontrun{
set.seed(1000)
alpha0 = 1.5
alpha1 = -0.3
beta1 = -0.61335306
n=200
crate =0
beta2=0
tau=1
gamma0 = -log(0.4)
maxE = 2
u = c(1.3, 1.8, 3)
E = runif(2*n, min=0, max=maxE)          # enrollment times
Z1 = (c(rep(0,n),rep(1,n)))                   # treatment indicator
Z2 = as.matrix(rnorm(2*n))                             # covariates
alpha = alpha0+alpha1*Z1
gamma1 = gamma0*exp(beta1*Z1+beta2*Z2)
FT = rweibull(2*n,shape=alpha,scale=gamma1**(-1/alpha))
CT = NULL
if (crate == 0)
{CT = Inf } else
{CT = rexp(2*n, rate=crate)}
Data = cbind(E,FT,CT,Z1,Z2)
delta = as.numeric(FT < CT)
X = pmin(FT,CT)
test <- gsRMST(tau = tau,Time = X,Status = delta,Z = Z2,TRT = Z1, Imax=1100, E = E, alpha = 0.05, u =u)
test$Z #1.051475 1.990320 2.232713
test$Crit # 2.818662 2.383526 2.022924
}




}
\references{
Zhang, P.K., Logan, B.L., and Martens, M.J. (2024). Covariate-adjusted Group Sequential Comparisons of Survival Probabilities. \emph{arXiv}

Zhang, X., Loberiza, F. R., Klein, J. P., and Zhang, M.-J. (2007). A SAS Macro for
Estimation of Direct Adjusted Survival Curves Based on a Stratified Cox Regression
Model. \emph{Comput Methods Programs Biomed} \strong{88(2)}, 95–101.

Zhang, X. (2013). Comparison of Restricted Mean Survival Times Between Treatments
Based on a Stratified Cox Model. \emph{Bio-Algorithms and Med-Systems} \strong{9(4)}, 183-189

Zucker, D.M. (1998) Restricted Mean Life with Covariates: Modification and Extension
of a Useful Survival Analysis Method. \emph{J Am Stat Assoc} \strong{93(442)}, 702-709
}
