% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspbinary.R
\name{Imaxaspbinary}
\alias{Imaxaspbinary}
\title{Title
Assume the event time for each subject follows a Weibull distribution with survival function
$S(t|Z_W, \mathbf{Z}) = \exp(-\gamma t^{\alpha})$,
where the shape and rate parameters $\alpha$ and $\gamma$ depend on $Z_W$ and $\mathbf{Z}$ through the forms
$\alpha = \alpha_0 + \alpha_1Z_W$ and
$\gamma = \gamma_0 \exp(\beta_WZ_W + \bfbeta^T \bfZ)$.}
\usage{
Imaxaspbinary(alpha0, alpha1, gamma0, beta2, crate, t0, maxE, n, effect, NN, p)
}
\arguments{
\item{alpha0}{- parameter to specify in Weibull model}

\item{alpha1}{- parameter to specify in Weibull model. alpha1 = 0 means there are proportional hazards; alpha1 != 0 means the proportional hazards assumption is violated}

\item{gamma0}{- parameter to specify in Weibull model}

\item{beta2}{-vector of coefficients for non-treatment group binary variables}

\item{crate}{- censoring rate, assumes an exponential distribution}

\item{t0}{- pre-specified time at which adjusted survival probabilities for each group are calculated}

\item{maxE}{- maximum enrollment time. Assumes uniform enrollment between [0,E]}

\item{n}{- sample size per group}

\item{effect}{- targeted effect size}

\item{NN}{- number of iterations}

\item{p}{- vector of probabilities for non-treatment group binary variables}
}
\description{
Title
Assume the event time for each subject follows a Weibull distribution with survival function
$S(t|Z_W, \mathbf{Z}) = \exp(-\gamma t^{\alpha})$,
where the shape and rate parameters $\alpha$ and $\gamma$ depend on $Z_W$ and $\mathbf{Z}$ through the forms
$\alpha = \alpha_0 + \alpha_1Z_W$ and
$\gamma = \gamma_0 \exp(\beta_WZ_W + \bfbeta^T \bfZ)$.
}
\examples{
\dontrun{
Imaxaspbinary(alpha0=1.5, alpha1=-1, gamma0=-log(0.4), beta2=0, crate=0, t0=1, maxE=2, n=200, effect= 0, NN = 10000,p=c(0.5, 0.3))
}

}
