% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmstbinary.R
\name{Imaxrmstbinary}
\alias{Imaxrmstbinary}
\title{Calculates \eqn{I_{max}} for adjusted restricted mean survival time difference for binary covariates}
\usage{
Imaxrmstbinary(
  alpha0,
  alpha1,
  gamma0,
  beta,
  crate,
  tau,
  maxE,
  n,
  effect,
  NN,
  p
)
}
\arguments{
\item{alpha0}{Parameter to specify in Weibull model. See Details for more information.}

\item{alpha1}{Parameter to specify in Weibull model. See Details for more information. \eqn{\alpha_1  = 0 } means there are proportional hazards; \eqn{\alpha_1 \neq 0 }  means the proportional hazards assumption is violated}

\item{gamma0}{Parameter to specify in Weibull model. See Details for more information.}

\item{beta}{Vector of coefficients for binary covariates}

\item{crate}{Censoring rate, assumes an exponential distribution}

\item{tau}{Pre-specified survival horizon time for adjusted restricted mean survival times}

\item{maxE}{Maximum enrollment time. Assumes uniform enrollment between [0, maxE]}

\item{n}{Sample size per group}

\item{effect}{Targeted effect size}

\item{NN}{Number of iterations used to calculate the maximum information}

\item{p}{Vector of probabilities for binary covariates}
}
\value{
Maximum information for a trial with the given parameters.
}
\description{
Calculates \eqn{I_{max}} for adjusted restricted mean survival time difference for binary covariates
}
\examples{
\dontrun{
set.seed(90)
 Imaxrmstbinary(alpha0=1.5, alpha1=-0.3, gamma0=-log(0.4), beta=0, crate=0, tau=1,
 maxE=200, n=200, effect=0.088, NN=100,p=c(0.5,0.3) )
}
}
\references{
Zhang, P.K., Logan, B.L., and Martens, M.J. (2024). Covariate-adjusted Group Sequential Comparisons of Survival Probabilities. \emph{arXiv}

Zhang, X., Loberiza, F. R., Klein, J. P., and Zhang, M.-J. (2007). A SAS Macro for
Estimation of Direct Adjusted Survival Curves Based on a Stratified Cox Regression
Model. \emph{Comput Methods Programs Biomed} \strong{88(2)}, 95–101.

Zhang, X. (2013). Comparison of Restricted Mean Survival Times Between Treatments
Based on a Stratified Cox Model. \emph{Bio-Algorithms and Med-Systems} \strong{9(4)}, 183-189

Zucker, D.M. (1998) Restricted Mean Life with Covariates: Modification and Extension
of a Useful Survival Analysis Method. \emph{J Am Stat Assoc} \strong{93(442)}, 702-709
}
