% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspnormal.R
\name{Nasp}
\alias{Nasp}
\title{Sample size calculation for a normal covariate}
\usage{
Nasp(
  alpha0,
  alpha1,
  gamma0,
  beta2,
  crate,
  t0,
  maxE,
  m,
  effect,
  NN,
  alpha,
  beta
)
}
\arguments{
\item{alpha0}{- Parameter to specify in Weibull model. See Details for more information.}

\item{alpha1}{- Parameter to specify in Weibull model. See Details for more information. \eqn{\alpha_1  = 0 } means there are proportional hazards; \eqn{\alpha_1 \neq 0 }  means the proportional hazards assumption is violated}

\item{gamma0}{- Parameter to specify in Weibull model. See Details for more information.}

\item{beta2}{- Coefficient of normal covariate}

\item{crate}{- Censoring rate, assumes an exponential distribution}

\item{t0}{- Pre-specified time at which adjusted SPs for each group are calculated}

\item{maxE}{- Maximum enrollment time. Assumes uniform enrollment between [0, E]}

\item{m}{- Sample size used to calculate the maximum information, Imax}

\item{effect}{- Targeted effect size}

\item{NN}{- Number of iterations used to calculate the maximum information}

\item{alpha}{- Targeted type I error rate}

\item{beta}{- Targeted type II error rate}
}
\value{
Sample size for a trial with the given parameters
}
\description{
Calculates the sample size given power and effect size via Monte Carlo simulation.
This function first calculates \eqn{V_{0}} and  \eqn{V_{effect}},
the variance in the control group and treatment group, respectively,
by calculating the maximum information for both groups using an intial sample size of m .
The sample size N is then calculated using the equation
\deqn{N = (z_{1-\alpha/2}*\sqrt(V_0*M) + z_{1-\beta}*\sqrt(V_{effect}*M))^2/effect^2}
}
\details{
See Details section in \code{\link{Imaxasp}} on how trial data are simulated.
}
\examples{
\dontrun{
 set.seed(1234)
 Nasp(alpha0 = 1.5, alpha1=-1, gamma0=-log(0.4), beta2=0, crate=0, t0=1, maxE=2, m=400,
  effect=0.14, NN=10000, alpha=0.05, beta = 0.2)
}
}
\references{
Zhang, P.K., Logan, B.L., and Martens, M.J. (2024). Covariate-adjusted Group Sequential Comparisons of Survival Probabilities. \emph{arXiv}

Zhang, X., Loberiza, F. R., Klein, J. P., and Zhang, M.-J. (2007). A SAS Macro for
Estimation of Direct Adjusted Survival Curves Based on a Stratified Cox Regression
Model. \emph{Comput Methods Programs Biomed} \strong{88(2)}, 95–101.

Zucker, D.M. (1998) Restricted Mean Life with Covariates: Modification and Extension
of a Useful Survival Analysis Method. \emph{J Am Stat Assoc} \strong{93(442)}, 702-709
}
