% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmst.R
\name{rmst}
\alias{rmst}
\title{Title}
\usage{
rmst(t0, Time, Status, Z, TRT)
}
\arguments{
\item{t0}{- time point of analysis}

\item{Time}{- Observed Times}

\item{Status}{- Censoring Status}

\item{Z}{- non-treatment groupcovariates}

\item{TRT}{- treatment group indicator}
}
\value{

}
\description{
Title
}
