% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{root}
\alias{root}
\title{Title}
\usage{
root(alpha0, alpha1, gamma0, beta2, t0, effect)
}
\arguments{
\item{alpha0}{-}

\item{alpha1}{-}

\item{gamma0}{-}

\item{beta2}{-}

\item{t0}{-}

\item{effect}{-}
}
\description{
Title
}
