% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{rootasp}
\alias{rootasp}
\title{Title}
\usage{
rootasp(alpha0, alpha1, gamma0, beta2, t0, effect)
}
\arguments{
\item{alpha0}{- parameter to specify in Weibull model}

\item{alpha1}{- parameter to specify in Weibull model. alpha1 = 0 means there are proportional hazards; alpha1 != 0 means the proportional hazards assumption is violated}

\item{gamma0}{- parameter to specify in Weibull model}

\item{beta2}{- vector of coefficients for non-treatment group binary variables}

\item{t0}{- pre-specified time at which adjusted survival probabilities for each group are calculated}

\item{effect}{- targeted effect size}
}
\description{
Title
}
\examples{
rootasp(alpha0 = 1.5, alpha1=-1, gamma0=-log(0.4), beta2 =0, t0=1, effect=0.140996938)
}
