% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary.R
\name{rootrmstbinary}
\alias{rootrmstbinary}
\title{Calculates \eqn{\beta_W} given a restricted mean survival time effect size}
\usage{
rootrmstbinary(alpha0, alpha1, gamma0, beta, tau, effect, p)
}
\arguments{
\item{alpha0}{Parameter to specify in Weibull model. See Details for more information.}

\item{alpha1}{Parameter to specify in Weibull model. See Details for more information. alpha1 = 0 means there are proportional hazards; alpha1 != 0 means the proportional hazards assumption is violated}

\item{gamma0}{Parameter to specify in Weibull model. See Details for more information.}

\item{beta}{vector of coefficients for non-treatment group binary variables}

\item{tau}{Pre-specified time at which adjusted restricted mean survival times for each group are calculated}

\item{effect}{Targeted effect size}

\item{p}{Vector of probabilities for binary covariates}
}
\value{
\eqn{\beta_W}, the coefficient for the treatment group variable
}
\description{
Calculates \eqn{\beta_W}, the coefficient for the treatment group variable, that corresponds to a given restricted mean survival time effect size
}
\examples{
rootrmstbinary(alpha0 = 1.5, alpha1=-0.3, gamma0=-log(0.4), beta =0, tau = 1,
 effect=0.1, p =c(0.2,0.5))
}
