% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rweib}
\alias{rweib}
\title{Title}
\usage{
rweib(n, shape, rate)
}
\arguments{
\item{n}{- sample size}

\item{shape}{- shape parameter for Weibull distribution}

\item{rate}{- rate parameter for Weibull distribution}
}
\description{
Title
}
\keyword{internal}
