% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asp.R
\name{asp}
\alias{asp}
\title{Adjusted SP}
\usage{
asp(t0, Time, Status, Z, TRT)
}
\arguments{
\item{t0}{- Pre-specified time point of analysis}

\item{Time}{- Observed times}

\item{Status}{- Event indicator (0 = censored, 1 = observed)}

\item{Z}{- Non-treatment group covariates}

\item{TRT}{- Treatment group indicator (0 = control, 1 = treatment)}
}
\value{
\itemize{
  \item SPD - Adjusted SP difference estimate
  \item SED - Standard error estimate of adjusted SP difference estimate
  \item S0 - Adjusted SP estimate for treatment group 0
  \item S1 - Adjusted SP estimate for treatment group 1
  \item SE0 - Standard error estimate of adjusted SP estimate for treatment group 0
  \item SE1 - Standard error estimate of adjusted SP estimate for treatment group 1
}
}
\description{
Calculates the adjusted SP difference between two treatment groups
at a pre-specified time point via a treatment-stratified Cox proportional hazards model.
Standard error estimates for the estimated difference are also calculated.
Adjusted SP estimates and standard error estimates
are also provided for each group.
}
\examples{
set.seed(1234)
t0 = 1
TRT = c(rep(0,5),rep(1,5))
Z = cbind(rnorm(10))
FT = c(0.30,  0.58,  0.41,  0.0333,  0.58,  0.10,  0.83,  2.45,  8.7, 17.1)
CT = c(0.20, 0.8, 0.73, 0.24, 0.5, 0.25, 1.3, 3.6, 10, 20.5)
X = pmin(FT,CT)
Status = as.numeric(FT <= CT)     & (X <= t0)
Time = pmin(X,t0)
output = asp(t0,Time,Status,Z,TRT)
output$SPD
output$SED
}
\references{
Zhang, P.K., Logan, B.L., and Martens, M.J. (2024). Covariate-adjusted Group Sequential Comparisons of Survival Probabilities. \emph{arXiv}

Zhang, X., Loberiza, F. R., Klein, J. P., and Zhang, M.-J. (2007). A SAS Macro for
Estimation of Direct Adjusted Survival Curves Based on a Stratified Cox Regression
Model. \emph{Comput Methods Programs Biomed} \strong{88(2)}, 95–101.

Zucker, D.M. (1998) Restricted Mean Life with Covariates: Modification and Extension
of a Useful Survival Analysis Method. \emph{J Am Stat Assoc} \strong{93(442)}, 702-709
}
