% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asp.R
\name{asp}
\alias{asp}
\title{asp}
\usage{
asp(t0, Time, Status, Z, TRT)
}
\arguments{
\item{t0}{- time point of analysis}

\item{Time}{- Observed Times (0 = Observed, 1 = Censored)}

\item{Status}{- Censoring Status (0 = Censored, 1 = Observed)}

\item{Z}{- Non-treatment groupcovariates}

\item{TRT}{- Treatment group indicator (0 = Control, 1 = Treatment)}
}
\value{
\itemize{
  \item SPD - restricted mean survival time differene estimate
  \item SE2 - standard error estimate of restricted mean survival time difference estimate
  \item S0 - adjusted survival probability of treatment group 0
  \item S1 - adjusted survival probability of treatment group 1
  \item se10 - standard error of adjusted survival probability of treatment group 0
  \item se11 - standard error of adjusted survival probability of treatment group 1
}
}
\description{
Calculates adjusted survival probability difference between two treatment groups
at a pre-specified time point via a treatment-stratified Cox proportional hazards model.
Variance and standard error estimates for the difference are also calculated.
Adjusted survival probability estimates and standard error / variance estimates
are also provided for each group.
}
\details{

}
\examples{
library(survival)
t0 = 1
TRT = c(rep(0,5),rep(1,5))
Z = cbind(rnorm(10))
FT = c(0.30,  0.58,  0.41,  0.0333,  0.58,  0.10,  0.83,  2.45,  8.7, 17.1)
CT = c(0.20, 0.8, 0.73, 0.24, 0.5, 0.25, 1.3, 3.6, 10, 20.5)
X = pmin(FT,CT)
Status = as.numeric(FT <= CT)     & (X <= t0)
Time = pmin(X,t0)
output = asp(t0,Time,Status,Z,TRT)
output$SPD
output$SE2
}
\references{

}
\seealso{
survival
}
\author{
Peter Zhang, Brent Logan, Michael Martens
}
