% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmstbinary.R
\name{ESrmstbinary}
\alias{ESrmstbinary}
\title{Title}
\usage{
ESrmstbinary(
  alpha0,
  alpha1,
  gamma0,
  beta2,
  crate,
  t0,
  maxE,
  n,
  NN,
  alpha = 0.05,
  beta = 0.2,
  max.iter = 10,
  p
)
}
\arguments{
\item{alpha0}{- parameter to specify in Weibull model. See Details for more information.}

\item{alpha1}{- parameter to specify in Weibull model. \eqn{\alpha_1  = 0 } means there are proportional hazards; \eqn{\alpha_1 \neq 0 }  means the proportional hazards assumption is violated}

\item{gamma0}{- parameter to specify in Weibull model. See Details for more information.}

\item{beta2}{-vector of coefficients for non-treatment group binary variables}

\item{crate}{- censoring rate, assumes an exponential distribution}

\item{t0}{- pre-specified time at which adjusted restricted mean survival times for each group are calculated}

\item{maxE}{- maximum enrollment time. Assumes uniform enrollment between [0,E]}

\item{n}{- sample size per group}

\item{NN}{- number of iterations}

\item{alpha}{- targeted type I error rate}

\item{p}{- vector of probabilities for non-treatment group binary variables}
}
\description{
Title
}
\examples{
\dontrun{}
}
