% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspbinary.R
\name{Imaxaspbinary}
\alias{Imaxaspbinary}
\title{Maximum information for adjusted survival probabilities for binary covariates}
\usage{
Imaxaspbinary(alpha0, alpha1, gamma0, beta2, crate, t0, maxE, n, effect, NN, p)
}
\arguments{
\item{alpha0}{- parameter to specify in Weibull model. See Details for more information.}

\item{alpha1}{- parameter to specify in Weibull model. \eqn{\alpha_1  = 0 } means there are proportional hazards; \eqn{\alpha_1 \neq 0 }  means the proportional hazards assumption is violated}

\item{gamma0}{- parameter to specify in Weibull model. See Details for more information.}

\item{beta2}{-vector of coefficients for non-treatment group binary variables}

\item{crate}{- censoring rate, assumes an exponential distribution}

\item{t0}{- pre-specified time at which adjusted survival probabilities for each group are calculated}

\item{maxE}{- maximum enrollment time. Assumes uniform enrollment between [0,E]}

\item{n}{- sample size per group}

\item{effect}{- targeted effect size}

\item{NN}{- number of iterations}

\item{p}{- vector of probabilities for non-treatment group binary variables}
}
\description{
Maximum information for adjusted survival probabilities for binary covariates
}
\details{
Trial data are simulated using the following assumptions.
Assume the event time for a subject follows a Weibull distribution with survival function
\deqn{S(t|Z_W, \mathbf{Z}) = \exp(-\gamma t^{\alpha}),}
where the shape and rate parameters
\eqn{\gamma} and \eqn{\alpha}  depend on \eqn{Z_W} (treatment group indicator)
and \eqn{\mathbf{Z}} (vector of baseline covariates):
 \deqn{\alpha = \alpha_0 + \alpha_1Z_W,}
\deqn{\gamma = \gamma_0 \exp(\beta_WZ_W + \mathbf{\beta}^T \mathbf{Z}).}
\eqn{\alpha_0},  \eqn{\alpha_1}, and \eqn{\gamma_0} are parameters that are chosen to
simulate trial data under a proportional hazards model, delayed effect setting, or crossing curves setting.
}
\examples{
\dontrun{
Imaxaspbinary(alpha0=1.5, alpha1=-1, gamma0=-log(0.4), beta2=0, crate=0, t0=1, maxE=2, n=200, effect= 0, NN = 10000,p=c(0.5, 0.3))
}
}
\references{
Zhang, P.K., Logan, B.L., and Martens, M.J. (2024). Covariate-adjusted Group Sequential Comparisons of Survival Probabilities. \emph{arXiv}

Zhang, X., Loberiza, F. R., Klein, J. P., and Zhang, M.-J. (2007). A SAS Macro for
Estimation of Direct Adjusted Survival Curves Based on a Stratified Cox Regression
model. \emph{Comput Methods Programs Biomed} \strong{88(2)}, 95–101.

Zucker, D.M. (1998) Restricted Mean Life with Covariates: Modification and Extension
of a Useful Survival Analysis Method. \emph{J Am Stat Assoc} \strong{93(442)}, 702-709
}
