% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmstbinary.R
\name{Imaxrmstbinary}
\alias{Imaxrmstbinary}
\title{Title}
\usage{
Imaxrmstbinary(
  alpha0,
  alpha1,
  gamma0,
  beta2,
  crate,
  t0,
  maxE,
  n,
  effect,
  NN,
  p
)
}
\arguments{
\item{alpha0}{- parameter to specify in Weibull model}

\item{alpha1}{- parameter to specify in Weibull model. alpha1 = 0 means there are proportional hazards; alpha1 != 0 means the proportional hazards assumption is violated}

\item{gamma0}{- parameter to specify in Weibull model}

\item{beta2}{-vector of coefficients for non-treatment group binary variables}

\item{crate}{- censoring rate, assumes an exponential distribution}

\item{t0}{- pre-specified time at which adjusted restricted mean survival times for each group are calculated}

\item{maxE}{- maximum enrollment time. Assumes uniform enrollment between [0,E]}

\item{n}{- sample size per group}

\item{effect}{- targeted effect size}

\item{NN}{- number of iterations}

\item{p}{- vector of probabilities for non-treatment group binary variables}
}
\description{
Title
}
