% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspnormal.R
\name{Nasp}
\alias{Nasp}
\title{Title}
\usage{
Nasp(
  alpha0,
  alpha1,
  gamma0,
  beta2,
  crate,
  t0,
  maxE,
  m,
  effect,
  NN,
  alpha,
  beta
)
}
\arguments{
\item{alpha0}{- parameter to specify in Weibull model}

\item{alpha1}{- parameter to specify in Weibull model. alpha1 = 0 means there are proportional hazards; alpha1 != 0 means the proportional hazards assumption is violated}

\item{gamma0}{- parameter to specify in Weibull model}

\item{beta2}{-vector of coefficients for non-treatment group binary variables}

\item{crate}{- censoring rate, assumes an exponential distribution}

\item{t0}{- pre-specified time at which adjusted survival probabilities for each group are calculated}

\item{maxE}{- maximum enrollment time. Assumes uniform enrollment between [0,E]}

\item{m}{- sample size used to calculate the maximum information, Imax}

\item{effect}{- targeted effect size}

\item{NN}{- number of iterations}

\item{alpha}{- targeted type I error rate}

\item{beta-}{targeted type II error rate}
}
\description{
Title
}
\examples{
\dontrun{
 Nasp(alpha0 = 1.5, alpha1=-1, gamma0=-log(0.4), beta2=0, crate=0, t0=1, maxE=2, m=400, effect=0.140996938, NN=10000, alpha=0.05, beta = 0.2)
}
}
