% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspbinary.R
\name{Naspbinary}
\alias{Naspbinary}
\title{Sample size calculation for testing survival probability difference adjusted for binary covariates}
\usage{
Naspbinary(
  alpha0,
  alpha1,
  gamma0,
  beta,
  crate,
  t0,
  maxE,
  m,
  effect,
  NN,
  alpha = 0.05,
  pi = 0.8,
  p
)
}
\arguments{
\item{alpha0}{Parameter to specify in Weibull model. See Details for more information.}

\item{alpha1}{Parameter to specify in Weibull model. See Details for more information. \eqn{\alpha_1  = 0 } means there are proportional hazards; \eqn{\alpha_1 \neq 0 }  means the proportional hazards assumption is violated}

\item{gamma0}{Parameter to specify in Weibull model. See Details for more information.}

\item{beta}{Vector of coefficients for binary covariates}

\item{crate}{Censoring rate, assumes an exponential distribution}

\item{t0}{Pre-specified time at which adjusted SPs for each group are calculated}

\item{maxE}{Maximum enrollment time. Assumes uniform enrollment between [0, maxE]}

\item{m}{Sample size used to calculate the maximum information, Imax}

\item{effect}{Targeted effect size}

\item{NN}{Number of iterations used to calculate the maximum information}

\item{alpha}{Targeted type I error rate}

\item{pi}{Targeted power}

\item{p}{Vector of probabilities for binary covariates}
}
\value{
Sample size for a trial with the given parameters
}
\description{
Sample size calculation for testing survival probability difference adjusted for binary covariates
}
\details{
Calculates the sample size for testing survival probability difference given power and effect size via Monte Carlo simulation.
 See Details section in \code{\link{Imaxasp}} on how trial data are simulated.
 By using MC replicate datasets with an initial sample size M, we can calculate
 \eqn{V_{0}^M=1/I_{0}^M} and \eqn{V_{effect}^M=1/I_{effect}^M}, which represent the variances
 of the estimated differences in M patient trials.
 Under the large sample normal distribution for the effect size estimator,
  the required sample size depends on the unit variance,
  which is the reciprocal of the amount of information contributed by a single patient.
  To obtain the unit variances under the null and targeted alternative hypotheses,
   \eqn{V_{0}^1} and \eqn{V_{effect}^1}, we rescale \eqn{V_{0}^M} and \eqn{V_{effect}^M} by calculating
\eqn{V_{0}^1 = V_{0}^M \cdot M } and \eqn{V_{effect}^1 = V_{effect}^M \cdot M }.
With the individual variances per patient, we can calculate the actual required total sample size N for the trial using the equation
\deqn{N = \frac{\left(z_{1-\alpha/2}\sqrt{V_0^1} + z_{\pi} \sqrt{V_{effect}^1}\right)^2}{effect^2},}
}
\examples{
\dontrun{
set.seed(8)
Naspbinary(alpha0 = 1.5, alpha1=-1, gamma0=-log(0.4), beta=log(2)/sqrt(2), crate=0, t0=1,
 maxE=2, m=500, effect=0.121, NN=10000, alpha=0.05, pi = 0.8, p =c(0.5, 0.3)) #192.27*2
}

}
\references{
Zhang, P.K., Logan, B.L., and Martens, M.J. (2024). Covariate-adjusted Group Sequential Comparisons of Survival Probabilities. \emph{arXiv}. https://arxiv.org/abs/2403.17117.

Zhang, X., Loberiza, F. R., Klein, J. P., and Zhang, M.-J. (2007). A SAS Macro for
Estimation of Direct Adjusted Survival Curves Based on a Stratified Cox Regression
Model. \emph{Comput Methods Programs Biomed} \strong{88(2)}, 95–101.

Zucker, D.M. (1998) Restricted Mean Life with Covariates: Modification and Extension
of a Useful Survival Analysis Method. \emph{J Am Stat Assoc} \strong{93(442)}, 702-709
}
