% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmstnormal.R
\name{Nrmst}
\alias{Nrmst}
\title{Sample size calculation for a normal covariate}
\usage{
Nrmst(
  alpha0,
  alpha1,
  gamma0,
  beta2,
  crate,
  t0,
  maxE,
  m,
  effect,
  NN,
  alpha,
  beta
)
}
\arguments{
\item{alpha0}{- Parameter to specify in Weibull model. See Details for more information.}

\item{alpha1}{- Parameter to specify in Weibull model. See Details for more information. \eqn{\alpha_1  = 0 } means there are proportional hazards; \eqn{\alpha_1 \neq 0 }  means the proportional hazards assumption is violated}

\item{gamma0}{- Parameter to specify in Weibull model. See Details for more information.}

\item{beta2}{- Coefficient of normal covariate}

\item{crate}{- Censoring rate, assumes an exponential distribution}

\item{t0}{- Pre-specified time at which adjusted RMSTs for each group are calculated}

\item{maxE}{- Maximum enrollment time. Assumes uniform enrollment between [0, E]}

\item{m}{- Sample size used to calculate the maximum information, Imax}

\item{effect}{- Targeted effect size}

\item{NN}{- Number of iterations used to calculate the maximum information}

\item{alpha}{- Targeted type I error rate}

\item{beta}{- Targeted type II error rate}
}
\value{
Power for a trial with the given parameters.
}
\description{
Sample size calculation for a normal covariate
}
\details{
See Details section in \code{\link{Imaxasp}} on how trial data are simulated.
}
\examples{
\dontrun{
set.seed(1234)
Nrmst(alpha0 = 1.5, alpha1 = -0.3, gamma0 = -log(0.4), crate = 0, t0=1,
maxE=2, m = 400, beta2=0, effect = 0.088, NN = 100, alpha = 0.05, beta=0.2)
}
}
\references{
Zhang, P.K., Logan, B.L., and Martens, M.J. (2024). Covariate-adjusted Group Sequential Comparisons of Survival Probabilities. \emph{arXiv}

Zhang, X., Loberiza, F. R., Klein, J. P., and Zhang, M.-J. (2007). A SAS Macro for
Estimation of Direct Adjusted Survival Curves Based on a Stratified Cox Regression
Model. \emph{Comput Methods Programs Biomed} \strong{88(2)}, 95–101.

Zhang, X. (2013). Comparison of Restricted Mean Survival Times Between Treatments
Based on a Stratified Cox Model. \emph{Bio-Algorithms and Med-Systems} \strong{9(4)}, 183-189

Zucker, D.M. (1998) Restricted Mean Life with Covariates: Modification and Extension
of a Useful Survival Analysis Method. \emph{J Am Stat Assoc} \strong{93(442)}, 702-709
}
