% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspnormal.R
\name{powerasp}
\alias{powerasp}
\title{Power calculation for a normal covariate}
\usage{
powerasp(alpha0, alpha1, gamma0, beta2, crate, t0, maxE, n, effect, NN, alpha)
}
\arguments{
\item{alpha0}{- Parameter to specify in Weibull model. See Details for more information.}

\item{alpha1}{- Parameter to specify in Weibull model. See Details for more information. \eqn{\alpha_1  = 0 } means there are proportional hazards; \eqn{\alpha_1 \neq 0 }  means the proportional hazards assumption is violated}

\item{gamma0}{- Parameter to specify in Weibull model. See Details for more information.}

\item{beta2}{- Coefficient of normal covariate}

\item{crate}{- Censoring rate, assumes an exponential distribution}

\item{t0}{- Pre-specified time at which adjusted SPs for each group are calculated}

\item{maxE}{- Maximum enrollment time. Assumes uniform enrollment between [0, E]}

\item{n}{- Sample size per group}

\item{effect}{- Targeted effect size}

\item{NN}{- Number of iterations used to calculate the maximum information}

\item{alpha}{- Targeted type I error rate}
}
\value{
Power for a trial with the given parameters.
}
\description{
Calculates the power given sample size and effect size via Monte Carlo simulation.
This function first calculates \eqn{V_{0}} and  \eqn{V_{effect}},
the variance in the control group and treatment group, respectively,
by calculating the maximum information for both groups.
The power is then calculated using the equation
\deqn{(\sqrt(N)*effect - z_{1-alpha/2}*\sqrt(V_0*N))/sqrt(V_{effect}*N)}

\eqn{I_{max}}, the maximum information for the trial, via Monte Carlo simulation.
This function simulates NN number of trials, calculates NN number of
adjusted SP differences between the treatment and control group, and
then takes the inverse of the variances of these differences to obtain \eqn{I_{max}}.
}
\details{
See Details section in \code{\link{Imaxasp}} on how trial data are simulated.
}
\examples{
\dontrun{
 powerasp(alpha0 = 1.5, alpha1= -1, gamma0=-log(0.4), beta2=0, crate=0, t0=1,
 maxE=2, n = 191, effect=0.14, NN=10000, alpha=0.05)
}
}
\references{
Zhang, P.K., Logan, B.L., and Martens, M.J. (2024). Covariate-adjusted Group Sequential Comparisons of Survival Probabilities. \emph{arXiv}

Zhang, X., Loberiza, F. R., Klein, J. P., and Zhang, M.-J. (2007). A SAS Macro for
Estimation of Direct Adjusted Survival Curves Based on a Stratified Cox Regression
Model. \emph{Comput Methods Programs Biomed} \strong{88(2)}, 95–101.

Zucker, D.M. (1998) Restricted Mean Life with Covariates: Modification and Extension
of a Useful Survival Analysis Method. \emph{J Am Stat Assoc} \strong{93(442)}, 702-709
}
