% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmst.R
\name{rmst}
\alias{rmst}
\title{RMST}
\usage{
rmst(t0, Time, Status, Z, TRT)
}
\arguments{
\item{t0}{- Pre-specified time point, RMST estimate is calculated over(0, t0)}

\item{Time}{- Observed times}

\item{Status}{- Event indicator (0 = Censored, 1 = Observed)}

\item{Z}{- Non-treatment group covariates}

\item{TRT}{- Treatment group indicator (0 = Control, 1 = Treatment)}
}
\value{
\itemize{
  \item muD - Adjusted SP difference estimate
  \item SED - Standard error estimate of adjusted SP difference estimate
  \item mu0 - Adjusted SP of treatment group 0
  \item mu1 - Adjusted SP of treatment group 1
  \item SE0 - Standard error of adjusted SP of treatment group 0
  \item SE1 - Standard error of adjusted SP of treatment group 1
}
}
\description{
Calculates RMST difference between two treatment groups
at a pre-specified time point via a treatment-stratified Cox proportional hazards model.
Standard error estimates for the estimated differences are also calculated.
RMST estimates and standard error estimates
are also provided for each group.
}
\examples{
set.seed(123)
t0 = 1
n0 = 100
n1 = 100
n = n0 + n1
alpha0 = 1.5
alpha1 = -0.3
gamma0 = -log(0.4)
beta1 = -0.5
beta2 = log(1.5)
crate = -log(0.95)
TRT = c(rep(0,n0),rep(1,n1))                        # treatment indicator
Z = cbind(rnorm(n))                               # covariates
alpha = alpha0+alpha1*TRT
gamma1 = gamma0*exp(beta1*TRT+beta2*Z)
FT = rweibull(n,shape=alpha,scale=gamma1**(-1/alpha))
CT = rexp(n, rate=crate)
X = pmin(FT,CT)
Status = as.numeric(FT <= CT)     & (X <= t0)
Time = pmin(X,t0)
rmst(t0,Time,Status,Z,TRT)$muD
rmst(t0,Time,Status,Z,TRT)$SED

}
\references{
Zhang, P.K., Logan, B.L., and Martens, M.J. (2024). Covariate-adjusted Group Sequential Comparisons of Survival Probabilities. \emph{arXiv}

Zhang, X., Loberiza, F. R., Klein, J. P., and Zhang, M.-J. (2007). A SAS Macro for
Estimation of Direct Adjusted Survival Curves Based on a Stratified Cox Regression
Model. \emph{Comput Methods Programs Biomed} \strong{88(2)}, 95–101.

Zhang, X. (2013). Comparison of Restricted Mean Survival Times Between Treatments
Based on a Stratified Cox Model. \emph{Bio-Algorithms and Med-Systems} \strong{9(4)}, 183-189

Zucker, D.M. (1998) Restricted Mean Life with Covariates: Modification and Extension
of a Useful Survival Analysis Method. \emph{J Am Stat Assoc} \strong{93(442)}, 702-709
}
