% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmst.R
\name{rmst}
\alias{rmst}
\title{Restricted Mean Survival Time}
\usage{
rmst(t0, Time, Status, Z, TRT)
}
\arguments{
\item{t0}{- pre-specified time point, rmst is calculated over(0, t0)}

\item{Time}{- Observed times}

\item{Status}{- Censoring indicator (0 = Censored, 1 = Observed)}

\item{Z}{- Non-treatment group covariates}

\item{TRT}{- Treatment group indicator (0 = Control, 1 = Treatment)}
}
\value{
\itemize{
  \item Delta - restricted mean survival time differene estimate
  \item DSE2 - standard error estimate of restricted mean survival time difference estimate
  \item mu0 - restricted mean survival time of treatment group 0
  \item mu1 - restricted mean survival time of treatment group 1
  \item se10 - standard error of restricted mean survival time of treatment group 0
  \item se11 - standard error of restricted mean survival time of treatment group 1
}
}
\description{
Calculates restricted mean survival time difference between two treatment groups
at a pre-specified time point via a treatment-stratified Cox proportional hazards model.
Variance and standard error estimates for the difference are also calculated.
Restricted mean survival time estimates and standard error / variance estimates
are also provided for each group.
}
\examples{
library(survival)
set.seed(123)
t0 = 1
n0 = 100
n1 = 100
n = n0 + n1
alpha0 = 1.5
alpha1 = -0.3
gamma0 = -log(0.4)
beta1 = -0.5
beta2 = log(1.5)
crate = -log(0.95)
TRT = c(rep(0,n0),rep(1,n1))                        # treatment indicator
Z = cbind(rnorm(n))                               # covariates
alpha = alpha0+alpha1*TRT
gamma1 = gamma0*exp(beta1*TRT+beta2*Z)
FT = rweibull(n,shape=alpha,scale=gamma1**(-1/alpha))
CT = rexp(n, rate=crate)
X = pmin(FT,CT)
Status = as.numeric(FT <= CT)     & (X <= t0)
Time = pmin(X,t0)
rmst(t0,Time,Status,Z,TRT)$Delta
rmst(t0,Time,Status,Z,TRT)$DSE2

}
\seealso{
survival
}
\author{
Peter Zhang, Brent Logan, Michael Martens
}
