% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{rootasp}
\alias{rootasp}
\title{Calculates \eqn{\beta_W} given a survival probability effect size}
\usage{
rootasp(alpha0, alpha1, gamma0, beta, t0, effect)
}
\arguments{
\item{alpha0}{Parameter to specify in Weibull model See Details for more information.}

\item{alpha1}{Parameter to specify in Weibull model. See Details for more information. alpha1 = 0 means there are proportional hazards; alpha1 != 0 means the proportional hazards assumption is violated}

\item{gamma0}{Parameter to specify in Weibull model. See Details for more information.}

\item{beta}{Vector of coefficients for non-treatment group binary variables}

\item{t0}{Pre-specified time at which adjusted SPs for each group are calculated}

\item{effect}{Targeted effect size}
}
\value{
\eqn{\beta_W}, the coefficient for the treatment group variable
}
\description{
Calculates \eqn{\beta_W}, the coefficient for the treatment group variable, that corresponds to a given survival probability effect size
}
\examples{
rootasp(alpha0 = 1.5, alpha1=-1, gamma0=-log(0.4), beta =0, t0=1,
effect=0.14)
}
